Refactor pattern fragment to use inheritance.
